+++ src/calendar.js
@@ -4,7 +4,7 @@
  * A javascript drag & drop event calendar, that is fully responsive and compatible with all modern browsers.
  * 
  * @file        calendar.js
- * @version     v2.0.6
+ * @version     v2.0.7
  * @author      Bunoon
  * @license     GNU AGPLv3
  * @copyright   Bunoon 2023
@@ -104,6 +104,7 @@
  * @property    {Object}    onEventDoubleClick                          Specifies an event that will be triggered when an event is double clicked (when editing mode is disabled, passes the event to the function).
  * @property    {Object}    onVisibleGroupsChanged                      Specifies an event that will be triggered when the visible groups are changed (passes the visible group names to the function).
  * @property    {Object}    onVisibleEventTypesChanged                  Specifies an event that will be triggered when the visible event types are changed (passes the visible event type IDs to the function).
+ * @property    {Object}    onNotification                              Specifies an event that will be triggered when a notification is shown (passes the event to the function).
  * 
  * These are the translatable strings that are used in Calendar.js.
  * 
@@ -579,6 +580,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
         _element_FullDayView_TodayButton = null,
         _element_FullDayView_TimeArrow = null,
         _element_FullDayView_SearchButton = null,
+        _element_FullDayView_ContextMenu_ClickPositionHourMinutes = null,
         _element_ListAllEventsView = null,
         _element_ListAllEventsView_ExportEventsButton = null,
         _element_ListAllEventsView_FullScreenButton = null,
@@ -646,6 +648,8 @@ function calendarJs( elementOrId, options, searchOptions ) {
         _element_DropDownMenu_Event_Copy = null,
         _element_DropDownMenu_Event_RemoveSeparator = null,
         _element_DropDownMenu_Event_Remove = null,
+        _element_DropDownMenu_Event_ExportEventsSeparator = null,
+        _element_DropDownMenu_Event_ExportEvents = null,
         _element_DropDownMenu_FullDay = null,
         _element_DropDownMenu_FullDay_RemoveEvents_Separator = null,
         _element_DropDownMenu_FullDay_RemoveEvents = null,
@@ -760,8 +764,8 @@ function calendarJs( elementOrId, options, searchOptions ) {
     function getAllEvents() {
         var events = [];
     
-        getAllEventsFunc( function( event ) {
-            events.push( event );
+        getAllEventsFunc( function( eventDetails ) {
+            events.push( eventDetails );
         } );
 
         return events;
@@ -810,10 +814,10 @@ function calendarJs( elementOrId, options, searchOptions ) {
         var onYesEvent = function() {
             onNoEvent();
 
-            getAllEventsFunc( function( event ) {
-                var repeatEvery = getNumber( event.repeatEvery );
-                if ( repeatEvery === _repeatType.never && compareFunc( event.from, date ) ) {
-                    _this.removeEvent( event.id, false );
+            getAllEventsFunc( function( eventDetails ) {
+                var repeatEvery = getNumber( eventDetails.repeatEvery );
+                if ( repeatEvery === _repeatType.never && compareFunc( eventDetails.from, date ) ) {
+                    _this.removeEvent( eventDetails.id, false );
                 }
             } );
 
@@ -1123,8 +1127,8 @@ function calendarJs( elementOrId, options, searchOptions ) {
         return className;
     }
 
-    function getAdjustedAllDayEvent( event ) {
-        var adjustedEvent = event;
+    function getAdjustedAllDayEvent( eventDetails ) {
+        var adjustedEvent = eventDetails;
 
         if ( adjustedEvent.isAllDay ) {
             adjustedEvent.from = new Date( adjustedEvent.from.getFullYear(), adjustedEvent.from.getMonth(), adjustedEvent.from.getDate(), 0, 0 );
@@ -1741,8 +1745,8 @@ function calendarJs( elementOrId, options, searchOptions ) {
         if ( !_datePickerModeEnabled ) {
             var yearsHandledForEvents = [];
 
-            getAllEventsFunc( function( event ) {
-                var fromYear = event.from.getFullYear();
+            getAllEventsFunc( function( eventDetails ) {
+                var fromYear = eventDetails.from.getFullYear();
     
                 if ( yearsHandledForEvents.indexOf( fromYear ) === -1 ) {
                     
@@ -2113,7 +2117,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
     }
 
     function isWorkingDay( date ) {
-        return _options.workingDays.indexOf( date.getDay() ) >= 0;
+        return _options.workingDays.indexOf( getWeekdayNumber( date ) ) >= 0;
     }
 
     function moveDateBackOneDay( date ) {
@@ -2676,6 +2680,10 @@ function calendarJs( elementOrId, options, searchOptions ) {
             _element_FullDayView.appendChild( _element_FullDayView_Contents );
     
             _element_FullDayView_Contents.oncontextmenu = function( e ) {
+                 var hoursMinutes = getHourMinutesFromMousePositionClick( e );
+
+                 _element_FullDayView_ContextMenu_ClickPositionHourMinutes = padNumber( hoursMinutes[ 0 ] ) + ":" + padNumber( hoursMinutes[ 1 ] );
+
                 showFullDayDropDownMenu( e );
             };
     
@@ -2710,11 +2718,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
 
     function fullDayViewDoubleClick( e ) {
         if ( _options.manualEditingEnabled ) {
-            var contentHoursOffset = getOffset( _element_FullDayView_Contents_Hours ),
-                scrollPosition = getScrollPosition(),
-                pixelsPerMinute = getFullDayPixelsPerMinute(),
-                minutesFromTop = Math.floor( ( e.pageY - ( contentHoursOffset.top + scrollPosition.top ) ) / pixelsPerMinute ),
-                hoursMinutes = getHoursAndMinutesFromMinutes( minutesFromTop );
+            var hoursMinutes = getHourMinutesFromMousePositionClick( e );
             
             if ( _options.useTemplateWhenAddingNewEvent ) {
                 var newBlankTemplateEventTime = padNumber( hoursMinutes[ 0 ] ) + ":" + padNumber( hoursMinutes[ 1 ] ),
@@ -2758,45 +2762,45 @@ function calendarJs( elementOrId, options, searchOptions ) {
             createSpanElement( _element_FullDayView_Title, " (" + holidayText + ")", "light-title-bar-text" );
         }
 
-        getAllEventsFunc( function( event ) {
-            var totalDays = getTotalDaysBetweenDates( event.from, event.to ) + 1,
-                nextDate = new Date( event.from );
+        getAllEventsFunc( function( eventDetails ) {
+            var totalDays = getTotalDaysBetweenDates( eventDetails.from, eventDetails.to ) + 1,
+                nextDate = new Date( eventDetails.from );
             
             for ( var dayIndex = 0; dayIndex < totalDays; dayIndex++ ) {
                 if ( doDatesMatch( nextDate, date ) ) {
-                    orderedEvents.push( event );
+                    orderedEvents.push( eventDetails );
                     break;
                 }
 
                 moveDateForwardDay( nextDate );
             }
             
-            var repeatEvery = getNumber( event.repeatEvery );
+            var repeatEvery = getNumber( eventDetails.repeatEvery );
             if ( repeatEvery > _repeatType.never ) {
                 if ( repeatEvery === _repeatType.everyDay ) {
-                    buildFullDayRepeatedDayEvents( event, orderedEvents, date, moveDateForwardDay, 1 );
+                    buildFullDayRepeatedDayEvents( eventDetails, orderedEvents, date, moveDateForwardDay, 1 );
                 } else if ( repeatEvery === _repeatType.everyWeek ) {
-                    buildFullDayRepeatedDayEvents( event, orderedEvents, date, moveDateForwardWeek, 1 );
+                    buildFullDayRepeatedDayEvents( eventDetails, orderedEvents, date, moveDateForwardWeek, 1 );
                 } else if ( repeatEvery === _repeatType.every2Weeks ) {
-                    buildFullDayRepeatedDayEvents( event, orderedEvents, date, moveDateForwardWeek, 2 );
+                    buildFullDayRepeatedDayEvents( eventDetails, orderedEvents, date, moveDateForwardWeek, 2 );
                 } else if ( repeatEvery === _repeatType.everyMonth ) {
-                    buildFullDayRepeatedDayEvents( event, orderedEvents, date, moveDateForwardMonth, 1 );
+                    buildFullDayRepeatedDayEvents( eventDetails, orderedEvents, date, moveDateForwardMonth, 1 );
                 } else if ( repeatEvery === _repeatType.everyYear ) {
-                    buildFullDayRepeatedDayEvents( event, orderedEvents, date, moveDateForwardYear, 1 );
+                    buildFullDayRepeatedDayEvents( eventDetails, orderedEvents, date, moveDateForwardYear, 1 );
                 } else if ( repeatEvery === _repeatType.custom ) {
 
-                    var repeatEveryCustomType = getNumber( event.repeatEveryCustomType ),
-                        repeatEveryCustomValue = getNumber( event.repeatEveryCustomValue );
+                    var repeatEveryCustomType = getNumber( eventDetails.repeatEveryCustomType ),
+                        repeatEveryCustomValue = getNumber( eventDetails.repeatEveryCustomValue );
                     
                     if ( repeatEveryCustomValue > 0 ) {
                         if ( repeatEveryCustomType === _repeatCustomType.daily ) {
-                            buildFullDayRepeatedDayEvents( event, orderedEvents, date, moveDateForwardDay, repeatEveryCustomValue );
+                            buildFullDayRepeatedDayEvents( eventDetails, orderedEvents, date, moveDateForwardDay, repeatEveryCustomValue );
                         } else if ( repeatEveryCustomType === _repeatCustomType.weekly ) {
-                            buildFullDayRepeatedDayEvents( event, orderedEvents, date, moveDateForwardWeek, repeatEveryCustomValue );
+                            buildFullDayRepeatedDayEvents( eventDetails, orderedEvents, date, moveDateForwardWeek, repeatEveryCustomValue );
                         } else if ( repeatEveryCustomType === _repeatCustomType.monthly ) {
-                            buildFullDayRepeatedDayEvents( event, orderedEvents, date, moveDateForwardMonth, repeatEveryCustomValue );
+                            buildFullDayRepeatedDayEvents( eventDetails, orderedEvents, date, moveDateForwardMonth, repeatEveryCustomValue );
                         } else if ( repeatEveryCustomType === _repeatCustomType.yearly ) {
-                            buildFullDayRepeatedDayEvents( event, orderedEvents, date, moveDateForwardYear, repeatEveryCustomValue );
+                            buildFullDayRepeatedDayEvents( eventDetails, orderedEvents, date, moveDateForwardYear, repeatEveryCustomValue );
                         }
                     }
                 }
@@ -2843,18 +2847,18 @@ function calendarJs( elementOrId, options, searchOptions ) {
         _element_FullDayView_EventsShown = [];
     }
 
-    function buildFullDayRepeatedDayEvents( event, orderedEvents, date, dateFunc, dateFuncForwardValue ) {
-        var newFromDate = new Date( event.from ),
-            excludeDays = getArray( event.repeatEveryExcludeDays );
+    function buildFullDayRepeatedDayEvents( eventDetails, orderedEvents, date, dateFunc, dateFuncForwardValue ) {
+        var newFromDate = new Date( eventDetails.from ),
+            excludeDays = getArray( eventDetails.repeatEveryExcludeDays );
     
         while ( newFromDate < date ) {
             dateFunc( newFromDate, dateFuncForwardValue );
 
-            var repeatEnded = !( !isDefined( event.repeatEnds ) || isDateSmallerOrEqualToDate( newFromDate, event.repeatEnds ) );
+            var repeatEnded = !( !isDefined( eventDetails.repeatEnds ) || isDateSmallerOrEqualToDate( newFromDate, eventDetails.repeatEnds ) );
 
             if ( excludeDays.indexOf( newFromDate.getDay() ) === -1 && !repeatEnded ) {
                 if ( doDatesMatch( newFromDate, date ) ) {
-                    orderedEvents.push( event );
+                    orderedEvents.push( eventDetails );
                     break;
                 }
             }
@@ -3078,6 +3082,16 @@ function calendarJs( elementOrId, options, searchOptions ) {
         }
 
         event.style.zIndex = zIndex.toString();
+    }
+
+    function getHourMinutesFromMousePositionClick( e ) {
+        var contentHoursOffset = getOffset( _element_FullDayView_Contents_Hours ),
+            scrollPosition = getScrollPosition(),
+            pixelsPerMinute = getFullDayPixelsPerMinute(),
+            minutesFromTop = Math.floor( ( e.pageY - ( contentHoursOffset.top + scrollPosition.top ) ) / pixelsPerMinute ),
+            hoursMinutes = getHoursAndMinutesFromMinutes( minutesFromTop );
+
+        return hoursMinutes;
     }
 
 
@@ -3752,7 +3766,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
     function buildListAllWeekEventsEvent( eventDetails, header, container, displayDate ) {
         var added = false,
             weekDayNumber = getWeekdayNumber( displayDate ),
-            dateID = displayDate.getDate() + displayDate.getFullYear() + displayDate.getMonth() + weekDayNumber,
+            dateID = weekDayNumber + displayDate.getMonth() + displayDate.getFullYear(),
             seriesIgnoreDates = getArray( eventDetails.seriesIgnoreDates ),
             formattedDate = toStorageFormattedDate( displayDate );
 
@@ -3875,7 +3889,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
 
     function buildListAllEventsDay( date ) {
         var weekDayNumber = getWeekdayNumber( date ),
-            dateID = date.getDate() + date.getFullYear() + date.getMonth() + weekDayNumber,
+            dateID = weekDayNumber + date.getMonth() + date.getFullYear(),
             dayContents = null,
             dayHeader = null,
             removeEventsDate = new Date( date );
@@ -4601,6 +4615,8 @@ function calendarJs( elementOrId, options, searchOptions ) {
             _element_DropDownMenu_Event_EditEvent = null;
             _element_DropDownMenu_Event_RemoveSeparator = null;
             _element_DropDownMenu_Event_Remove = null;
+            _element_DropDownMenu_Event_ExportEventsSeparator = null;
+            _element_DropDownMenu_Event_ExportEvents = null;
         }
 
         _element_DropDownMenu_Event = createElement( "div", "calendar-drop-down-menu" );
@@ -4683,6 +4699,14 @@ function calendarJs( elementOrId, options, searchOptions ) {
         _element_DropDownMenu_Event_OpenUrl = buildMenuItemWithIcon( _element_DropDownMenu_Event, "ib-arrow-top-right-icon", _options.openUrlText, function() {
             openEventUrl( _element_DropDownMenu_Event_EventDetails.url );
         } );
+
+        if ( _options.exportEventsEnabled ) {
+            _element_DropDownMenu_Event_ExportEventsSeparator = buildMenuSeparator( _element_DropDownMenu_Event );
+
+            _element_DropDownMenu_Event_ExportEvents = buildMenuItemWithIcon( _element_DropDownMenu_Event, "ib-arrow-down-full-line-icon", _options.exportEventsTooltipText + "...", function() {
+                showSelectExportTypeDialog( _eventsSelected );
+            } );
+        }
     }
 
     function buildFullDayViewDropDownMenu() {
@@ -4702,12 +4726,12 @@ function calendarJs( elementOrId, options, searchOptions ) {
 
             buildMenuItemWithIcon( _element_DropDownMenu_FullDay, "ib-plus-icon", _options.addEventTitle + "...", function() {
                 if ( _options.useTemplateWhenAddingNewEvent ) {
-                    var newBlankTemplateEvent = buildBlankTemplateEvent( _element_FullDayView_DateSelected, _element_FullDayView_DateSelected );
+                    var newBlankTemplateEvent = buildBlankTemplateEvent( _element_FullDayView_DateSelected, _element_FullDayView_DateSelected, _element_FullDayView_ContextMenu_ClickPositionHourMinutes, _element_FullDayView_ContextMenu_ClickPositionHourMinutes );
 
                     showEventEditingDialog( newBlankTemplateEvent );
                     showEventEditingDialogTitleSelected();
                 } else {
-                    showEventEditingDialog( null, _element_FullDayView_DateSelected );
+                    showEventEditingDialog( null, _element_FullDayView_DateSelected, _element_FullDayView_ContextMenu_ClickPositionHourMinutes );
                 }
             }, true );
 
@@ -4836,6 +4860,11 @@ function calendarJs( elementOrId, options, searchOptions ) {
                 _element_DropDownMenu_Event_OpenUrlSeparator.style.display = "none";
                 _element_DropDownMenu_Event_OpenUrl.style.display = "none";
 
+                if ( _options.exportEventsEnabled ) {
+                    _element_DropDownMenu_Event_ExportEventsSeparator.style.display = "block";
+                    _element_DropDownMenu_Event_ExportEvents.style.display = "block";
+                }
+
             } else if ( locked ) {
                 if ( _options.manualEditingEnabled ) {
                     _element_DropDownMenu_Event_EditEvent.style.display = "block";
@@ -4861,6 +4890,11 @@ function calendarJs( elementOrId, options, searchOptions ) {
                     _element_DropDownMenu_Event_OpenUrl.style.display = "none";
                 }
 
+                if ( _options.exportEventsEnabled ) {
+                    _element_DropDownMenu_Event_ExportEventsSeparator.style.display = "none";
+                    _element_DropDownMenu_Event_ExportEvents.style.display = "none";
+                }
+
             } else {
                 if ( _options.manualEditingEnabled ) {
                     _element_DropDownMenu_Event_EditEvent.style.display = "block";
@@ -4884,6 +4918,11 @@ function calendarJs( elementOrId, options, searchOptions ) {
                     _element_DropDownMenu_Event_OpenUrl.style.display = "block";
                 } else {
                     _element_DropDownMenu_Event_OpenUrl.style.display = "none";
+                }
+
+                if ( _options.exportEventsEnabled ) {
+                    _element_DropDownMenu_Event_ExportEventsSeparator.style.display = "none";
+                    _element_DropDownMenu_Event_ExportEvents.style.display = "none";
                 }
             }
 
@@ -5030,12 +5069,12 @@ function calendarJs( elementOrId, options, searchOptions ) {
             _element_EventEditorDialog_ErrorMessage = createElement( "p", "error" );
             contents.appendChild( _element_EventEditorDialog_ErrorMessage );
     
-            var buttonsSplitContainer = createElement( "div", "split" );
-            contents.appendChild( buttonsSplitContainer );
+            var buttonsContainer = createElement( "div", "buttons-container" );
+            contents.appendChild( buttonsContainer );
     
-            _element_EventEditorDialog_OKButton = createButtonElement( buttonsSplitContainer, _options.addText, "ok", eventDialogEvent_OK );
-            createButtonElement( buttonsSplitContainer, _options.cancelText, "cancel", eventDialogEvent_Cancel );
-            _element_EventEditorDialog_RemoveButton = createButtonElement( contents, _options.removeEventText, "remove", eventDialogEvent_Remove );
+            _element_EventEditorDialog_RemoveButton = createButtonElement( buttonsContainer, _options.removeEventText, "remove", eventDialogEvent_Remove );
+            _element_EventEditorDialog_OKButton = createButtonElement( buttonsContainer, _options.addText, "ok", eventDialogEvent_OK );
+            createButtonElement( buttonsContainer, _options.cancelText, "cancel", eventDialogEvent_Cancel );
         }
     }
 
@@ -5257,7 +5296,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             setEventTypeInputCheckedStates( eventDetails.type );
 
             _element_EventEditorDialog_OKButton.value = _options.updateText;
-            _element_EventEditorDialog_RemoveButton.style.display = "block";
+            _element_EventEditorDialog_RemoveButton.style.display = "inline-block";
             _element_EventEditorDialog_EventDetails = eventDetails;
             _element_EventEditorDialog_TimeFrom.value = toFormattedTime( eventDetails.from );
             _element_EventEditorDialog_TimeTo.value = toFormattedTime( eventDetails.to );
@@ -5660,11 +5699,11 @@ function calendarJs( elementOrId, options, searchOptions ) {
     
             setInputType( _element_EventEditorColorsDialog_ColorBorder, "color" );
     
-            var buttonsSplitContainer = createElement( "div", "split" );
-            contents.appendChild( buttonsSplitContainer );
+            var buttonsContainer = createElement( "div", "buttons-container" );
+            contents.appendChild( buttonsContainer );
     
-            createButtonElement( buttonsSplitContainer, _options.okText, "ok", eventColorsDialogEvent_OK );
-            createButtonElement( buttonsSplitContainer, _options.cancelText, "cancel", eventColorsDialogEvent_Cancel );
+            createButtonElement( buttonsContainer, _options.okText, "ok", eventColorsDialogEvent_OK );
+            createButtonElement( buttonsContainer, _options.cancelText, "cancel", eventColorsDialogEvent_Cancel );
         }
     }
 
@@ -5729,11 +5768,11 @@ function calendarJs( elementOrId, options, searchOptions ) {
     
             setInputType( _element_EventEditorRepeatOptionsDialog_RepeatEnds, "date" );
     
-            var buttonsSplitContainer = createElement( "div", "split" );
-            contents.appendChild( buttonsSplitContainer );
+            var buttonsContainer = createElement( "div", "buttons-container" );
+            contents.appendChild( buttonsContainer );
     
-            createButtonElement( buttonsSplitContainer, _options.okText, "ok", eventRepeatOptionsDialogEvent_OK );
-            createButtonElement( buttonsSplitContainer, _options.cancelText, "cancel", eventRepeatOptionsDialogEvent_Cancel );
+            createButtonElement( buttonsContainer, _options.okText, "ok", eventRepeatOptionsDialogEvent_OK );
+            createButtonElement( buttonsContainer, _options.cancelText, "cancel", eventRepeatOptionsDialogEvent_Cancel );
         }
     }
 
@@ -5811,16 +5850,16 @@ function calendarJs( elementOrId, options, searchOptions ) {
             _element_ConfirmationDialog_RemoveAllEvents = checkbox[ 0 ];
             _element_ConfirmationDialog_RemoveAllEvents_Label = checkbox[ 1 ];
     
-            var buttonsSplitContainer = createElement( "div", "split" );
-            contents.appendChild( buttonsSplitContainer );
-    
-            _element_ConfirmationDialog_YesButton = createElement( "input", "yes", "button" );
-            _element_ConfirmationDialog_YesButton.value = _options.yesText;
-            buttonsSplitContainer.appendChild( _element_ConfirmationDialog_YesButton );
+            var buttonsContainer = createElement( "div", "buttons-container" );
+            contents.appendChild( buttonsContainer );
     
             _element_ConfirmationDialog_NoButton = createElement( "input", "no", "button" );
             _element_ConfirmationDialog_NoButton.value = _options.noText;
-            buttonsSplitContainer.appendChild( _element_ConfirmationDialog_NoButton );
+            buttonsContainer.appendChild( _element_ConfirmationDialog_NoButton );
+
+            _element_ConfirmationDialog_YesButton = createElement( "input", "yes", "button" );
+            _element_ConfirmationDialog_YesButton.value = _options.yesText;
+            buttonsContainer.appendChild( _element_ConfirmationDialog_YesButton );
         }
     }
 
@@ -5897,11 +5936,11 @@ function calendarJs( elementOrId, options, searchOptions ) {
             _element_SelectExportTypeDialog_Option_HTML = buildRadioButton( radioButtonsContainer2, "HTML", "ExportType" );
             _element_SelectExportTypeDialog_Option_TSV = buildRadioButton( radioButtonsContainer2, "TSV", "ExportType" );
     
-            var buttonsSplitContainer = createElement( "div", "split" );
-            contents.appendChild( buttonsSplitContainer );
+            var buttonsContainer = createElement( "div", "buttons-container" );
+            contents.appendChild( buttonsContainer );
     
-            createButtonElement( buttonsSplitContainer, _options.okText, "ok", exportEventsFromOptionSelected );
-            createButtonElement( buttonsSplitContainer, _options.cancelText, "cancel", hideSelectExportTypeDialog );
+            createButtonElement( buttonsContainer, _options.okText, "ok", exportEventsFromOptionSelected );
+            createButtonElement( buttonsContainer, _options.cancelText, "cancel", hideSelectExportTypeDialog );
         }
     }
 
@@ -6061,11 +6100,11 @@ function calendarJs( elementOrId, options, searchOptions ) {
     
             _element_SearchDialog_Option_Contains.checked = true;
     
-            var buttonsSplitContainer = createElement( "div", "split last-split" );
-            _element_SearchDialog_Contents.appendChild( buttonsSplitContainer );
+            var buttonsContainer = createElement( "div", "buttons-container" );
+            _element_SearchDialog_Contents.appendChild( buttonsContainer );
     
-            _element_SearchDialog_Previous = createButtonElement( buttonsSplitContainer, _options.previousText, "previous", searchOnPrevious );
-            _element_SearchDialog_Next = createButtonElement( buttonsSplitContainer, _options.nextText, "next", searchOnNext );
+            _element_SearchDialog_Previous = createButtonElement( buttonsContainer, _options.previousText, "previous", searchOnPrevious );
+            _element_SearchDialog_Next = createButtonElement( buttonsContainer, _options.nextText, "next", searchOnNext );
         }
     }
 
@@ -6218,14 +6257,14 @@ function calendarJs( elementOrId, options, searchOptions ) {
             storeSearchOptions( true );
 
             for ( var orderedEventIndex = 0; orderedEventIndex < orderedEventsLength; orderedEventIndex++ ) {
-                var event = orderedEvents[ orderedEventIndex ];
+                var eventDetails = orderedEvents[ orderedEventIndex ];
 
-                if ( isEventVisible( event ) ) {
-                    var title = getString( event.title ),
-                        location = getString( event.location ),
-                        description = getString( event.description ),
-                        group = getString( event.group ),
-                        url = getString( event.url ),
+                if ( isEventVisible( eventDetails ) ) {
+                    var title = getString( eventDetails.title ),
+                        location = getString( eventDetails.location ),
+                        description = getString( eventDetails.description ),
+                        group = getString( eventDetails.group ),
+                        url = getString( eventDetails.url ),
                         found = false;
 
                     if ( !matchCase ) {
@@ -6253,17 +6292,17 @@ function calendarJs( elementOrId, options, searchOptions ) {
                     }
 
                     if ( found ) {
-                        var eventElement = getElementByID( startingID + event.id );
+                        var eventElement = getElementByID( startingID + eventDetails.id );
                         if ( eventElement !== null ) {
 
                             if ( isFullDayViewVisible || isAllEventsViewVisible || isAllWeekEventsViewVisible ) {
-                                _element_SearchDialog_SearchResults.push( event );
+                                _element_SearchDialog_SearchResults.push( eventDetails );
                             } else {
                                 
-                                var monthYear = event.from.getMonth() + "-" + event.from.getFullYear();
+                                var monthYear = eventDetails.from.getMonth() + "-" + eventDetails.from.getFullYear();
         
                                 if ( !monthYearsFound.hasOwnProperty( monthYear ) ) {
-                                    _element_SearchDialog_SearchResults.push( event );
+                                    _element_SearchDialog_SearchResults.push( eventDetails );
                                     monthYearsFound[ monthYear ] = true;
                                 }
                             }
@@ -6279,11 +6318,11 @@ function calendarJs( elementOrId, options, searchOptions ) {
         updateSearchButtons();
 
         if ( _element_SearchDialog_SearchResults.length > 0 ) {
-            var eventDetails = _element_SearchDialog_SearchResults[ _element_SearchDialog_SearchIndex ],
-                dateFrom = new Date( eventDetails.from );
+            var eventDetailsSearchResult = _element_SearchDialog_SearchResults[ _element_SearchDialog_SearchIndex ],
+                dateFrom = new Date( eventDetailsSearchResult.from );
 
             build( dateFrom );
-            updatedFocusedElementAfterSearch( eventDetails );
+            updatedFocusedElementAfterSearch( eventDetailsSearchResult );
         }
     }
 
@@ -6583,11 +6622,11 @@ function calendarJs( elementOrId, options, searchOptions ) {
             _element_ConfigurationDialog_Organizer_Email = createElement( "input", null, "email" );
             _element_ConfigurationDialog_Organizer.appendChild( _element_ConfigurationDialog_Organizer_Email );
     
-            var buttonsSplitContainer = createElement( "div", "split" );
-            contents.appendChild( buttonsSplitContainer );
+            var buttonsContainer = createElement( "div", "buttons-container" );
+            contents.appendChild( buttonsContainer );
     
-            createButtonElement( buttonsSplitContainer, _options.okText, "ok", configurationDialogEvent_OK );
-            createButtonElement( buttonsSplitContainer, _options.cancelText, "cancel", configurationDialogEvent_Cancel );
+            createButtonElement( buttonsContainer, _options.okText, "ok", configurationDialogEvent_OK );
+            createButtonElement( buttonsContainer, _options.cancelText, "cancel", configurationDialogEvent_Cancel );
         }
     }
 
@@ -6643,10 +6682,10 @@ function calendarJs( elementOrId, options, searchOptions ) {
         _element_ConfigurationDialog.style.display = "none";
     }
 
-    function isEventVisible( event ) {
-        var group = getString( event.group ),
+    function isEventVisible( eventDetails ) {
+        var group = getString( eventDetails.group ),
             configGroup = getGroupName( group ),
-            type = getNumber( event.type ),
+            type = getNumber( eventDetails.type ),
             visible = true;
         
         if ( group !== _string.empty ) {
@@ -7054,51 +7093,51 @@ function calendarJs( elementOrId, options, searchOptions ) {
      * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
      */
 
-    function checkEventForBrowserNotifications( date, event ) {
+    function checkEventForBrowserNotifications( date, eventDetails ) {
         runBrowserNotificationAction( function() {
-            if ( isDateToday( date ) && !_eventNotificationsTriggered.hasOwnProperty( event.id ) ) {
-                if ( !isDefinedBoolean( event.showAlerts ) || event.showAlerts ) {
+            if ( isDateToday( date ) && !_eventNotificationsTriggered.hasOwnProperty( eventDetails.id ) ) {
+                if ( !isDefinedBoolean( eventDetails.showAlerts ) || eventDetails.showAlerts ) {
                     var newFrom = new Date(),
                         newTo = new Date(),
                         today = new Date(),
-                        repeatEvery = getNumber( event.repeatEvery );
+                        repeatEvery = getNumber( eventDetails.repeatEvery );
         
-                    newFrom.setHours( event.from.getHours(), event.from.getMinutes(), 0, 0 );
-                    newTo.setHours( event.to.getHours(), event.to.getMinutes(), 0, 0 );
+                    newFrom.setHours( eventDetails.from.getHours(), eventDetails.from.getMinutes(), 0, 0 );
+                    newTo.setHours( eventDetails.to.getHours(), eventDetails.to.getMinutes(), 0, 0 );
         
-                    if ( repeatEvery === _repeatType.never && !isDateToday( event.from ) ) {
+                    if ( repeatEvery === _repeatType.never && !isDateToday( eventDetails.from ) ) {
                         newFrom.setHours( 0, 0, 0, 0 );
                     }
         
-                    if ( repeatEvery === _repeatType.never && !isDateToday( event.to ) ) {
+                    if ( repeatEvery === _repeatType.never && !isDateToday( eventDetails.to ) ) {
                         newTo.setHours( 23, 59, 59, 99 );
                     }
                     
                     if ( today >= newFrom && today <= newTo ) {
-                        launchBrowserNotificationForEvent( event );
+                        launchBrowserNotificationForEvent( eventDetails );
                     }
                 }
             }
-        }, false );
+        }, false, eventDetails );
     }
 
-    function launchBrowserNotificationForEvent( event ) {
-        _eventNotificationsTriggered[ event.id ] = true;
+    function launchBrowserNotificationForEvent( eventDetails ) {
+        _eventNotificationsTriggered[ eventDetails.id ] = true;
     
         var notification = new Notification( _options.eventNotificationTitle, {
-            body: _options.eventNotificationBody.replace( "{0}", event.title ),
+            body: _options.eventNotificationBody.replace( "{0}", eventDetails.title ),
         });
 
         notification.onclick = function() {
-            var url = getString( event.url );
+            var url = getString( eventDetails.url );
 
             if ( url === _string.empty ) {
-                showEventEditingDialog( event );
+                showEventEditingDialog( eventDetails );
             } else {
                 openEventUrl( url );
             }
 
-            triggerOptionsEventWithData( "onNotificationClicked", event );
+            triggerOptionsEventWithData( "onNotificationClicked", eventDetails );
         };
     }
 
@@ -7110,7 +7149,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
         } );
     }
 
-    function runBrowserNotificationAction( action, writeConsoleLog ) {
+    function runBrowserNotificationAction( action, writeConsoleLog, eventDetails ) {
         if ( _options.eventNotificationsEnabled && !_datePickerModeEnabled ) {
             writeConsoleLog = isDefined( writeConsoleLog ) ? writeConsoleLog : true;
 
@@ -7120,6 +7159,10 @@ function calendarJs( elementOrId, options, searchOptions ) {
                 }
             } else {
                 action();
+            }
+
+            if ( isDefined( eventDetails ) ) {
+                triggerOptionsEventWithData( "onNotification", eventDetails );
             }
         }
     }
@@ -7362,8 +7405,8 @@ function calendarJs( elementOrId, options, searchOptions ) {
         var groups = [],
             groupsAnyCase = [];
 
-        getAllEventsFunc( function( event ) {
-            var group = getString( event.group );
+        getAllEventsFunc( function( eventDetails ) {
+            var group = getString( eventDetails.group );
             if ( group !== _string.empty && groupsAnyCase.indexOf( group.toLowerCase() ) === -1 ) {
                 groups.push( group );
                 groupsAnyCase.push( group.toLowerCase() );
@@ -8260,11 +8303,11 @@ function calendarJs( elementOrId, options, searchOptions ) {
         return eventContents;
     }
 
-    function getOrderedEventPropertyNameList( event ) {
+    function getOrderedEventPropertyNameList( eventDetails ) {
         var propertyNames = [];
 
-        for ( var propertyName in event ) {
-            if ( event.hasOwnProperty( propertyName ) ) {
+        for ( var propertyName in eventDetails ) {
+            if ( eventDetails.hasOwnProperty( propertyName ) ) {
                 propertyNames.push( propertyName );
             }
         }
@@ -9300,15 +9343,15 @@ function calendarJs( elementOrId, options, searchOptions ) {
             updateEvents = !isDefinedBoolean( updateEvents ) ? true : updateEvents;
             triggerEvent = !isDefinedBoolean( triggerEvent ) ? true : triggerEvent;
 
-            getAllEventsFunc( function( event ) {
-                if ( event.id === id ) {
-                    event.from = from;
-                    event.to = to;
-                    event.repeatEnds = repeatEnds;
+            getAllEventsFunc( function( eventDetails ) {
+                if ( eventDetails.id === id ) {
+                    eventDetails.from = from;
+                    eventDetails.to = to;
+                    eventDetails.repeatEnds = repeatEnds;
                     updated = true;
     
                     if ( triggerEvent ) {
-                        triggerOptionsEventWithData( "onEventUpdated", event );
+                        triggerOptionsEventWithData( "onEventUpdated", eventDetails );
                     }
     
                     if ( updateEvents ) {
@@ -9433,9 +9476,9 @@ function calendarJs( elementOrId, options, searchOptions ) {
         var returnEvent = null;
 
         if ( isDefinedString( id ) && !_datePickerModeEnabled ) {
-            getAllEventsFunc( function( event ) {
-                if ( event.id === id ) {
-                    returnEvent = event;
+            getAllEventsFunc( function( eventDetails ) {
+                if ( eventDetails.id === id ) {
+                    returnEvent = eventDetails;
                     return true;
                 }
             } );
@@ -9460,10 +9503,10 @@ function calendarJs( elementOrId, options, searchOptions ) {
             updateEvents = !isDefinedBoolean( updateEvents ) ? true : updateEvents;
             triggerEvent = !isDefinedBoolean( triggerEvent ) ? true : triggerEvent;
     
-            getAllEventsFunc( function( event ) {
-                var repeatEvery = getNumber( event.repeatEvery );
-                if ( repeatEvery === _repeatType.never && event.to < new Date() ) {
-                    _this.removeEvent( event.id, false, triggerEvent );
+            getAllEventsFunc( function( eventDetails ) {
+                var repeatEvery = getNumber( eventDetails.repeatEvery );
+                if ( repeatEvery === _repeatType.never && eventDetails.to < new Date() ) {
+                    _this.removeEvent( eventDetails.id, false, triggerEvent );
                 }
             } );
     
@@ -9612,8 +9655,8 @@ function calendarJs( elementOrId, options, searchOptions ) {
             updateEvents = !isDefinedBoolean( updateEvents ) ? true : updateEvents;
             triggerEvent = !isDefinedBoolean( triggerEvent ) ? true : triggerEvent;
 
-            getAllEventsFunc( function( event ) {
-                event.group = null;
+            getAllEventsFunc( function( eventDetails ) {
+                eventDetails.group = null;
             } );
 
             if ( triggerEvent ) {
@@ -9647,9 +9690,9 @@ function calendarJs( elementOrId, options, searchOptions ) {
 
             var checkGroupName = groupName.toLowerCase();
 
-            getAllEventsFunc( function( event ) {
-                if ( event.group !== null && event.group.toLowerCase() === checkGroupName ) {
-                    event.group = null;
+            getAllEventsFunc( function( eventDetails ) {
+                if ( eventDetails.group !== null && eventDetails.group.toLowerCase() === checkGroupName ) {
+                    eventDetails.group = null;
                 }
             } );
 
@@ -9792,7 +9835,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
      * @returns     {string}                                                The version number.
      */
     this.getVersion = function() {
-        return "2.0.6";
+        return "2.0.7";
     };
 
     /**